Bunmp deps and refactor getEthereuUrl
